+++ multipicker/src/main/java/im/vector/lib/multipicker/AudioPicker.kt
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.multipicker
+package im.vector.lib.multipicker
 
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.media.MediaMetadataRetriever
 import android.provider.MediaStore
-import im.vector.riotx.multipicker.entity.MultiPickerAudioType
+import im.vector.lib.multipicker.entity.MultiPickerAudioType
 
 /**
  * Audio file picker implementation