+++ multipicker/src/main/java/im/vector/lib/multipicker/FilePicker.kt
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.multipicker
+package im.vector.lib.multipicker
 
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.provider.OpenableColumns
-import im.vector.riotx.multipicker.entity.MultiPickerFileType
+import im.vector.lib.multipicker.entity.MultiPickerFileType
 
 /**
  * Implementation of selecting any type of files