+++ multipicker/src/main/java/im/vector/lib/multipicker/MultiPicker.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.multipicker
+package im.vector.lib.multipicker
 
 class MultiPicker<T> {