+++ multipicker/src/main/java/im/vector/lib/multipicker/entity/MultiPickerAudioType.kt
@@ -14,11 +14,14 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.multipicker.entity
+package im.vector.lib.multipicker.entity
 
-data class MultiPickerContactType(
-        val displayName: String,
-        val photoUri: String?,
-        val phoneNumberList: List<String>,
-        val emailList: List<String>
-)
+import android.net.Uri
+
+data class MultiPickerAudioType(
+        override val displayName: String?,
+        override val size: Long,
+        override val mimeType: String?,
+        override val contentUri: Uri,
+        val duration: Long
+) : MultiPickerBaseType