+++ multipicker/src/main/java/im/vector/lib/multipicker/entity/MultiPickerBaseType.kt
@@ -14,16 +14,13 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.multipicker.entity
+package im.vector.lib.multipicker.entity
 
 import android.net.Uri
 
-data class MultiPickerImageType(
-        override val displayName: String?,
-        override val size: Long,
-        override val mimeType: String?,
-        override val contentUri: Uri,
-        val width: Int,
-        val height: Int,
-        val orientation: Int
-) : MultiPickerBaseType
+interface MultiPickerBaseType {
+    val displayName: String?
+    val size: Long
+    val mimeType: String?
+    val contentUri: Uri
+}