+++ multipicker/src/main/java/im/vector/lib/multipicker/provider/MultiPickerFileProvider.kt
@@ -14,17 +14,8 @@
  * limitations under the License.
  */
 
-package im.vector.riotx.multipicker.entity
+package im.vector.lib.multipicker.provider
 
-import android.net.Uri
+import androidx.core.content.FileProvider
 
-data class MultiPickerVideoType(
-        override val displayName: String?,
-        override val size: Long,
-        override val mimeType: String?,
-        override val contentUri: Uri,
-        val width: Int,
-        val height: Int,
-        val orientation: Int,
-        val duration: Long
-) : MultiPickerBaseType
+class MultiPickerFileProvider : FileProvider()