+++ CHANGES.md
@@ -22,6 +22,7 @@ Build 🧱:
 Other changes:
  - Hide Flair settings, this is not implemented yet.
  - Rename package `im.vector.riotx.attachmentviewer` to `im.vector.lib.attachmentviewer`
+ - Rename package `im.vector.riotx.multipicker` to `im.vector.lib.multipicker`
 
 Changes in Element 1.0.4 (2020-08-03)
 ===================================================